Function to draw ROI + extracted Depth 1st frame
Created a function in order to draw our ROI easier. This ROI is going
to serve as premarked place where our object/s is/are going to be.

Saved and exctracted our 1st frame Depth as a csv file so we can see
what objects have been moved in the future frames.

if 1st_depth != actual_depth:
	robot doesn't track anymore that object